Used Local() instead location_factory()
