+++ tests/files/locations/test_local.py
@@ -6,7 +6,8 @@ import uuid
 import pytest
 
 from astro.constants import FileLocation
-from astro.files.locations import get_location_type, location_factory
+from astro.files.locations.base import Location
+from astro.files.locations.local import Local
 
 CWD = pathlib.Path(__file__).parent
 
@@ -35,14 +36,14 @@ def describe_get_location_type():
 
     def with_supported_location():  # skipcq: PTC-W0065, PYL-W0612
         """With all the supported locations"""
-        location = get_location_type(LOCAL_FILEPATH)
+        location = Location.get_location_type(LOCAL_FILEPATH)
         assert location == FileLocation.LOCAL
 
     def with_unsupported_location_raises_exception():  # skipcq: PYL-W0612, PTC-W0065
         """With all the unsupported locations, should raise a valueError exception"""
         unsupported_filepath = "invalid://some-file"
         with pytest.raises(ValueError) as exc_info:
-            _ = get_location_type(unsupported_filepath)
+            _ = Location.get_location_type(unsupported_filepath)
         expected_msg = "Unsupported scheme 'invalid' from path 'invalid://some-file'"
         assert exc_info.value.args[0] == expected_msg
 
@@ -52,7 +53,7 @@ def describe_get_transport_params():
 
     def with_local():  # skipcq: PYL-W0612, PTC-W0065
         """with local filepath"""
-        location = location_factory(LOCAL_FILEPATH)
+        location = Local(LOCAL_FILEPATH)
         credentials = location.get_transport_params()
         assert credentials is None
 
@@ -60,22 +61,18 @@ def describe_get_transport_params():
 def describe_get_paths():
     """test get_paths() method"""
 
-    def with_local_dir(local_dir):  # skipcq: PYL-W0612, PTC-W0065
+    @pytest.mark.parametrize(
+        "path", [LOCAL_DIR, LOCAL_DIR + "file_*"], ids=["without-prefix", "with-prefix"]
+    )
+    def with_local_dir(local_dir, path):  # skipcq: PYL-W0612, PTC-W0065
         """with local filepath"""
-        path = LOCAL_DIR
-        location = location_factory(path)
-        assert sorted(location.get_paths()) == [LOCAL_DIR_FILE_1, LOCAL_DIR_FILE_2]
-
-    def with_local_prefix(local_dir):  # skipcq: PYL-W0612, PTC-W0065
-        """with local filepath having glob pattern"""
-        path = LOCAL_DIR + "file_*"
-        location = location_factory(path)
+        location = Local(path)
         assert sorted(location.get_paths()) == [LOCAL_DIR_FILE_1, LOCAL_DIR_FILE_2]
 
     def with_unsupported_location(local_dir):  # skipcq: PYL-W0612, PTC-W0065
         """with unsupported filepath"""
         path = "invalid://some-file"
         with pytest.raises(ValueError) as exc_info:
-            _ = location_factory(path)
-        expected_msg = "Unsupported scheme 'invalid' from path 'invalid://some-file'"
+            _ = Local(path)
+        expected_msg = "Invalid path: 'invalid://some-file'"
         assert exc_info.value.args[0] == expected_msg
